Move api pagination core tests
