+++ tests/settings.py
@@ -52,4 +52,4 @@ PATTERNS_IGNORED_IN_QUERY_CAPTURES: List[Union[Pattern, SimpleLazyObject]] = [
     lazy_re_compile(r"^SET\s+")
 ]
 
-INSTALLED_APPS.append("tests.api.pagination")  # noqa: F405
+INSTALLED_APPS.append("saleor.tests")  # noqa: F405
